+++ besu/src/main/java/org/hyperledger/besu/cli/BesuCommand.java
@@ -1904,7 +1904,9 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
       if (unstableChainDataPruningOptions.getChainDataPruningBlocksRetained()
           < ChainDataPruningOptions.DEFAULT_CHAIN_DATA_PRUNING_MIN_BLOCKS_RETAINED) {
         throw new ParameterException(
-            this.commandLine, "--Xchain-data-pruning-blocks-retained must be >= 50400");
+            this.commandLine,
+            "--Xchain-data-pruning-blocks-retained must be >= "
+                + ChainDataPruningOptions.DEFAULT_CHAIN_DATA_PRUNING_MIN_BLOCKS_RETAINED);
       }
       if (unstableChainDataPruningOptions.getChainDataPruningBlocksFrequency() < 0) {
         throw new ParameterException(